ARROW-11575: [Developer][Archery] Move benchmarks codec to benchmarks dir
